Add the initial writing/saving logic
This is still in a testing stage. The current logic will look for a 'modded.gsc' file located in the application folder to overwrite the GSC contents of the save file. Later I will change the richtext editor to a syntax highlighted code editor and use that as the GSC to overwrite with rather than reading from a file.